fix(frontend): restore extOf and fileMentionRegex behavior after language module rename
The C2 P2 rename of lib/languages.ts → lib/language.ts silently regressed
three behavioural guarantees that downstream tests (lib/languages.test.ts)
and consumers (lib/markdown.ts inline file-mention parser, FileMention.tsx)
depended on:

- extOf(".gitignore") returned "gitignore" — restored to "" (a leading
  dotfile has no extension).
- extOf("foo.dir/bar") returned "dir" — restored to "" (when the last /
  is after the last ., the trailing segment is a directory).
- fileMentionRegex() greedily consumed surrounding characters, producing
  matches like "see notes/x.md" (whitespace), "(notes/x.md" (punctuation),
  and "://example.com/x.md" (URL body). Restored the lookbehind/lookahead
  regex so the match is exactly the path; URLs and absolute paths are
  correctly rejected. Also memoised the compiled regex.

Updated lib/languages.test.ts (the now-orphaned test file) to import from
./language so tsc --noEmit is clean. Added a CHANGELOG/unreleased/current.md
entry under "Unclassified" documenting the regressions and the fix.

Verification: tsc --noEmit clean; vitest run 177 files / 1461 tests pass
(incl. languages.test.ts 9/9, language.test.ts 5/5, markdown.test.ts
25/25, FileMention.test.tsx 3/3); go build ./... and go vet ./... clean
(frontend-only change).
